Rename CallInProgress.method to CallInProgress.http_method, tag v3
The method named "method" was giving me some grief, so I renamed it to
http_method, which is probably a more accurate name anyhow. As this is
a breaking API change, this is now v3.0.0 of the gem.

Also updated the test dependencies to Rails 3.2.12, since 3.2.8 is
pretty out of date.